backend/pyproject.toml
[tool.poetry]
name = "ragbot"
version = "0.0.1"
description = "A Retrieval Augmented Generation powered Chatbot"
authors = ["Anurag <19914954+shenoy-anurag@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.9"
fastapi = { extras = ["standard"], version = "^0.115.11" }
requests = "^2.32.3"
sqlmodel = "^0.0.24"
pydantic-settings = "^2.8.1"
pyjwt = "^2.10.1"
alembic = "^1.15.1"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
bcrypt = "^4.3.0"
apscheduler = "^3.11.0"
langchain = "^0.3.21"
langchain-openai = "^0.3.9"
langchain-qdrant = "^0.2.0"
langchain-community = "^0.3.20"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pandas = "^2.2.3"
nltk = "^3.9.1"
beautifulsoup4 = "^4.13.3"
ruff = "^0.9.10"
coverage = "^7.6.12"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
